--- a/packages/slate/src/models/node.js
+++ b/packages/slate/src/models/node.js
@@ -1627,7 +1627,7 @@ class Node {
 
   getTextsAtRange(range) {
     range = range.normalize(this)
-    if (range.isUnset) return []
+    if (range.isUnset) return List()
     const { startKey, endKey } = range
     return new List(this.getTextsBetweenPositionsAsArray(startKey, endKey))
   }
@@ -1665,19 +1665,8 @@ class Node {
   getTextsAtRangeAsArray(range) {
     range = range.normalize(this)
     if (range.isUnset) return []
-
     const { startKey, endKey } = range
-    const startText = this.getDescendant(startKey)
-
-    // PERF: the most common case is when the range is in a single text node,
-    // where we can avoid a lot of iterating of the tree.
-    if (startKey == endKey) return [startText]
-
-    const endText = this.getDescendant(endKey)
-    const texts = this.getTextsAsArray()
-    const start = texts.indexOf(startText)
-    const end = texts.indexOf(endText)
-    return texts.slice(start, end + 1)
+    return this.getTextsBetweenPositionsAsArray(startKey, endKey)
   }
 
   /**