--- a/packages/slate/src/models/node.js
+++ b/packages/slate/src/models/node.js
@@ -1176,9 +1176,9 @@ class Node {
       if (this.getClosestBlock(key) !== this.getClosestBlock(previous.key)) {
         return OrderedSet()
       }
+
       const char = previous.characters.last()
       if (!char) return OrderedSet()
-
       return new OrderedSet(char.marks)
     }