This PR is a part of #1771 .  In this PR, we optimize the marks related methods, by 

## Optimization
1. because getCharacters builds a large list, taking  much time and heap memory.  Therefore, we try to avoid calling getCharacters from node.
2. AtRange functions caches are not very useful, because range can have same value but different reference, we use (At/Between)Positions functions instead.
3. rather than calling concat in reduce function, we use only one concat in the end.

## Code clean
For getInsertMarksAtRange, asArray function is unnecessary, because this function can read marks directly from char.