@zhujinxuan 2018-05-02
```
  benchmarks
    changes
      delete-backward
        user : 373.59 → 398.89 ops/sec
        hr : 666.03 → 709.01 ops/sec
      delete-forward
        user : 122.86 → 116.78 ops/sec
        hr : 176.81 → 189.03 ops/sec
      insert-text-by-key-multiple
        user : 73.46 → 69.27 ops/sec
        hr : 101.37 → 91.79 ops/sec
      insert-text-by-key
        user : 411.50 → 360.03 ops/sec
        hr : 688.16 → 603.70 ops/sec
      insert-text
        user : 1867.95 → 1769.93 ops/sec
        hr : 3037.85 → 3043.54 ops/sec
      normalize
        user : 649.97 → 631.54 ops/sec
        hr : 1036.24 → 1068.82 ops/sec
      split-block
        user : 406.65 → 419.25 ops/sec
        hr : 671.30 → 685.70 ops/sec
    models
      from-json-big
        user : 4.10 → 4.10 ops/sec
        hr : 5.60 → 5.51 ops/sec
      from-json
        user : 181.12 → 144.52 ops/sec
        hr : 297.90 → 194.05 ops/sec (54% slower)
      get-active-marks-at-range
        user : 353.93 → 4666.23 ops/sec (92% faster)
        hr : 462.45 → 6890.98 ops/sec (93% faster)
      get-blocks-at-range
        user : 659.83 → 780.68 ops/sec
        hr : 1125.66 → 1329.68 ops/sec
      get-blocks
        user : 7926.57 → 6842.24 ops/sec
        hr : 11934.08 → 10776.17 ops/sec
      get-characters-at-range
        user : 232.35 → 1239.72 ops/sec (81% faster)
        hr : 307.24 → 1824.07 ops/sec (83% faster)
      get-characters
        user : 3741.39 → 2665.67 ops/sec (40% slower)
        hr : 5053.89 → 3431.95 ops/sec (47% slower)
      get-inlines-at-range
        user : 546.47 → 628.03 ops/sec
        hr : 698.00 → 809.60 ops/sec
      get-inlines
        user : 11178.14 → 9441.53 ops/sec
        hr : 14269.92 → 14417.95 ops/sec
      get-leaves
        user : 8118.39 → 7585.95 ops/sec
        hr : 12409.74 → 13552.90 ops/sec
      get-marks-at-range
        user : 174.21 → NaN ops/sec
        hr : 231.29 → NaN ops/sec
      get-marks
        user : 641.18 → 1584.49 ops/sec (60% faster)
        hr : 891.35 → 2270.78 ops/sec (61% faster)
      get-path
        user : 1967.97 → 2013.69 ops/sec
        hr : 2778.14 → 2737.37 ops/sec
      get-texts-at-range
        user : 3912.20 → 3944.62 ops/sec
        hr : 5598.94 → 5401.18 ops/sec
      get-texts
        user : 9888.00 → 10597.11 ops/sec
        hr : 15258.48 → 15412.77 ops/sec
      has-node-multiple
        user : 2167.56 → 2028.93 ops/sec
        hr : 2484.41 → 2397.54 ops/sec
      has-node
        user : 9414.25 → 8761.46 ops/sec
        hr : 12104.39 → 12090.27 ops/sec
      to-json
        user : 706.60 → 614.60 ops/sec
        hr : 1133.68 → 1044.50 ops/sec
      update-node
        user : 2723.39 → 2952.74 ops/sec
        hr : 4179.84 → 4426.08 ops/sec
    html-serializer
      deserialize
        user : 8.58 → 8.86 ops/sec
        hr : 11.26 → 11.72 ops/sec
      serialize
        user : 250.70 → 245.21 ops/sec
        hr : 414.38 → 399.25 ops/sec
    plain-serializer
      deserialize
        user : 424.73 → 442.21 ops/sec
        hr : 638.39 → 718.39 ops/sec
      serialize
        user : 2235.47 → 2432.66 ops/sec
        hr : 3433.75 → 3494.50 ops/sec
    rendering
      normal
        user : 73.31 → 62.22 ops/sec
        hr : 101.66 → 65.05 ops/sec (56% slower)

✨  Done in 50.72s.
```