@zhujinxuan 2018-05-02
@ianstormtaylor  Hi, I have split a part of PR #1771 and ready for review.
Method to run the benchmark:
1. clone branch https://github.com/ianstormtaylor/slate/pull/1765 at your local git as benchmark-consistence
2. run `git checkout benchmark-consistence && lerna bootstrap`
3. run `yarn build && yarn benchmark:save` to have the benchmark baseline
3. checkout to this branch; `git checkout getActiveMarks && yarn build && git checkout benchmark-consistence && yarn benchmark`