@zhujinxuan 2018-05-02
Just found a bug in calculating the percentage in benchmark.  For example `438.54 → 4184.49 ops/sec (90% faster)`

Benchmark compare after fix:
```

  benchmarks
    changes
      delete-backward
        user : 453.27 → 468.48 ops/sec
        hr : 724.23 → 775.80 ops/sec
      delete-forward
        user : 120.95 → 118.14 ops/sec
        hr : 205.96 → 196.87 ops/sec
      insert-text-by-key-multiple
        user : 74.66 → 75.13 ops/sec
        hr : 104.43 → 105.30 ops/sec
      insert-text-by-key
        user : 398.83 → 503.91 ops/sec
        hr : 694.56 → 758.33 ops/sec
      insert-text
        user : 1738.79 → 1835.63 ops/sec
        hr : 2915.49 → 3125.09 ops/sec
      normalize
        user : 674.78 → 641.20 ops/sec
        hr : 1182.63 → 1093.15 ops/sec
      split-block
        user : 433.41 → 441.42 ops/sec
        hr : 691.84 → 715.17 ops/sec
    models
      from-json-big
        user : 4.36 → 4.38 ops/sec
        hr : 6.20 → 6.07 ops/sec
      from-json
        user : 191.16 → 168.86 ops/sec
        hr : 332.56 → 306.16 ops/sec
      get-active-marks-at-range
        user : 438.54 → 4184.49 ops/sec (854.1899240462549% faster) 🙌
        hr : 580.99 → 5858.20 ops/sec (908.320948454062% faster) 🙌
      get-blocks-at-range
        user : 610.68 → 681.10 ops/sec
        hr : 1230.07 → 1215.10 ops/sec
      get-blocks
        user : 7932.68 → 6993.46 ops/sec
        hr : 12012.98 → 10479.54 ops/sec
      get-characters-at-range
        user : 264.33 → 1309.22 ops/sec (395.2914652801355% faster) 🙌
        hr : 327.27 → 2074.73 ops/sec (533.9580011601297% faster) 🙌
      get-characters
        user : 4129.72 → 3184.90 ops/sec
        hr : 5421.76 → 4203.37 ops/sec
      get-inlines-at-range
        user : 596.31 → 534.98 ops/sec
        hr : 818.31 → 709.93 ops/sec
      get-inlines
        user : 11615.01 → 10340.44 ops/sec
        hr : 16744.22 → 15554.53 ops/sec
      get-leaves
        user : 6912.18 → 9261.00 ops/sec (33.98095350743401% faster)
        hr : 13879.51 → 16013.20 ops/sec
      get-marks-at-range
        user : 187.52 → 563.22 ops/sec (200.35764573359617% faster) 🙌
        hr : 250.25 → 931.57 ops/sec (272.25962550065987% faster) 🙌
      get-marks
        user : 651.59 → 1520.08 ops/sec (133.28994503357464% faster) 🙌
        hr : 985.22 → 2110.31 ops/sec (114.19618833908712% faster) 🙌
      get-path
        user : 1999.02 → 2166.49 ops/sec
        hr : 2998.23 → 2920.33 ops/sec
      get-texts-at-range
        user : 3822.58 → 3505.75 ops/sec
        hr : 5814.02 → 5391.79 ops/sec
      get-texts
        user : 9485.88 → 10475.31 ops/sec
        hr : 15565.11 → 16037.79 ops/sec
      has-node-multiple
        user : 2124.88 → 2156.52 ops/sec
        hr : 2490.17 → 2556.09 ops/sec
      has-node
        user : 7796.19 → 8896.54 ops/sec
        hr : 11686.07 → 12753.46 ops/sec
      to-json
        user : 687.10 → 685.10 ops/sec
        hr : 1083.30 → 1125.44 ops/sec
      update-node
        user : 3167.10 → 3196.03 ops/sec
        hr : 4764.91 → 4803.60 ops/sec
    html-serializer
      deserialize
        user : 9.16 → 9.12 ops/sec
        hr : 12.30 → 12.41 ops/sec
      serialize
        user : 256.04 → 251.17 ops/sec
        hr : 424.29 → 405.76 ops/sec
    plain-serializer
      deserialize
        user : 451.42 → 451.58 ops/sec
        hr : 742.07 → 737.24 ops/sec
      serialize
        user : 2483.58 → 2270.99 ops/sec
        hr : 3530.57 → 3448.53 ops/sec
    rendering
      normal
        user : 71.37 → 75.31 ops/sec
        hr : 102.30 → 103.82 ops/sec
```