@zhujinxuan 2018-05-03
@ianstormtaylor Do you have any further changes about this PR?  Then I can work on leaves/characters based on this PR in another branch.