@ianstormtaylor 2018-05-11
Thanks for taking the time to do this one @zhujinxuan! I see a handful of stylistic issues, but I'll fix them after merging, so that you're not blocked.